requirements.txt
Flask==3.1.0
Pillow==11.1.0
